Address code review comments
- Split compute kernels into separate 2 kernels
- Remove composite operation from SwiGLU forward op
- Fix typos in comments
- Link TODO comments with corresponding GitHub issues
- Add 2 additional test cases for better coverage
