+++ tt-train/sources/ttml/metal/ops/swiglu_fw/device/kernels/dataflow/writer_swiglu_fw_interleaved_start_id.cpp
@@ -2,8 +2,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <cmath>
-
 #include "dataflow_api.h"
 
 // CBs with input data
@@ -25,7 +23,7 @@ constexpr uint32_t block_size = get_compile_time_arg_val(0);
 constexpr uint32_t Wt = get_compile_time_arg_val(1);
 
 // TODO(maciek): Move all write_cb_block to common utils file, and reuse them in other
-// operations.
+// operations. See tracking issue #31125 for more details.
 template <typename AddrGen>
 inline void write_cb_block_to_dram(
     uint32_t cb_idx, const AddrGen& addr_gen, uint32_t start_idx, uint32_t current_block_size, uint32_t tile_bytes) {
