+++ tt-train/sources/ttml/metal/ops/swiglu_fw/device/swiglu_fw_program_factory.hpp
@@ -15,8 +15,8 @@ struct SwiGLUForwardProgramFactory {
         tt::tt_metal::KernelHandle swiglu_fw_writer_kernel_id;
         tt::tt_metal::KernelHandle swiglu_fw_kernel_group_1_id;
         tt::tt_metal::KernelHandle swiglu_fw_kernel_group_2_id;
-        CoreRangeSet core_group_1;
-        CoreRangeSet core_group_2;
+        tt::tt_metal::CoreRangeSet core_group_1;
+        tt::tt_metal::CoreRangeSet core_group_2;
         uint32_t num_cores{};
         uint32_t num_cores_y{};
     };
