+++ tt-train/sources/ttml/ops/swiglu_op.hpp
@@ -12,7 +12,6 @@ autograd::TensorPtr swiglu(
     const autograd::TensorPtr& tensor,
     const autograd::TensorPtr& w1,
     const autograd::TensorPtr& w2,
-    const autograd::TensorPtr& w3,
-    bool use_composite_fw = false);
+    const autograd::TensorPtr& w3);
 
 }  // namespace ttml::ops
